mgt/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
requests
celery
Authlib
